serve/requirements.txt
python-dotenv==1.0.0
pydantic>=2.5.0,<3.0.0
pandas>=2.2.0
numpy>=1.26.0
google-cloud-bigquery>=3.23.1
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
db-dtypes>=1.1.0
streamlit>=1.32.0
plotly>=5.20.0
altair>=5.2.0
